Removed parsers package in favor of the model package
