+++ tatlin/lib/model/stl/parser.py
